+++ source/extensions/omni.isaac.benchmarks/docs/CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## [0.2.0] - 2022-10-24
+
+### Added
+- Lidar Benchmark
+
+### Fixed
+- Camera Benchmark and logging bugs
+
+
 ## [0.1.0] - 2022-10-05
 
 ### Added